Drop LagomBundleKeys.endpointsPort
`LagomBundleKeys.endpointsPort` is deprecated. However, we set a default for this key in `LagomBundlePlugin`. Therefore for every Lagom project we currently get a warning that this key is depracted.

Instead we can remove this key because this change will be released as the version 2.0.0. This indicates that this is a breaking change.