Introduces the install command
A new command is provided that conveniently installs all of the bundles and their associated configurations (If present) to the local sandbox cluster. In addition, if the project is a Lagom one then the Cassandra configuration will be generated and loaded along with Cassandra.

A generateInstallationScript command generates an installation script that is encouraged to be copied and used as the basis of a script that can manage a collection of bundles, in particular a collection that represents a system of micro services as is the case with Lagom.