Corrected memory declarations
BundleKeys.memory was always intended to represent the amount of resident memory required for a bundle to run. In Mesos and with other container environments, this value could be enforced via cgroups or ulimit and so forth. The mistake made though was that BundleKeys.memory was being interpreted as heap memory size.

This PR now makes the strong distinction between JVM heap requirements (expressed as JavaOptions) and resident memory requirements.