Modify integration tests to make Travis happy
Starting ConductR 2.0.0 with the new sandbox takes longer than with ConductR 1.x. Therefore, running the integration tests on Travis are failing because they take longer than 50 minutes which is the maximum build time on Travis. After 50 minutes, Travis is stopping the build and marks it as failing.

This PR removes some integration and tests and merges tests so that the integrations tests on Travis take less than 50 minutes.

This PR also upgrades the Lagom version used in the tests to 1.2.2.