Remove cassandra configuration directories
This PR removes:
- The cassandra configuration directory `cassandra-conf` from the custom bundle configuration
- The `runtime-config.sh` from the custom bundle configuration

Both are not necessary with the latest `conductr-cassandra` bundle. The custom bundle configuration, that is produced with the `cassandra-configuration:dist` command on each Lagom project, only need to contain a `bundle.conf`.

As a result, we do not need to maintain the cassandra configuration files in `conductr-cassandra` and `sbt-conductr` whenever we bump the cassandra version.

Tested with the newest `conductr-cassandra` bundle in the sandbox.