Using conductr-bundle-lib 1.5.0
Reverting the conductr-bundle-lib version to 1.5.0 to support Lagom 1.2.x and Lagom 1.1.x projects.

After this is merged, I’ll create a sbt-conductr 2.2.5 version that can be used with Lagom 1.2.x and Lagom 1.3.x. Aditionally, I’ll create a sbt-conductr release 2.3.0-rc.1 that uses conductr-bundle-lib 1.3.0 and is therefore compatible with Lagom 1.3.x.