Remove Cassandra configuration
Removes the `cassandra-configuration` scopes. Since PR https://github.com/typesafehub/conductr-cassandra/pull/12 it is not necessary for a Lagom project to create a custom Cassandra bundle configuration.

The `install` command has been changed accordingly so that it does not uses the `cassandra-configuration:dist` command for a Lagom project.

Note that with this change the cassandra endpoints are always declared in the services, not the ACL format. This is the current endpoint format used in the default Cassandra bundle. Prior to this change the `cassandra-configuration:dist` command checked the `BundleKeys.enableAcls` key and created the endpoints respectively. Once we change the default Cassandra bundle configuration to the ACL endpoint format, it will not support ConductR 1.x anymore.