Use latest conductr-cli release
Currently, on Travis we install the master state of the conductr-cli. This can lead to errors in case that the latest commit has errors or is not compatible with released ConductR versions. See https://github.com/typesafehub/sbt-conductr/pull/226#issuecomment-285321013

This PR is installing the latest conductr-cli release on Travis instead. Also, for the other Prod-Suite components we are using the latest release instead of a snapshot version, e.g. ConductR, cinnamon, etc.