Travis: Use conductr-cli master branch
Reverting PR https://github.com/typesafehub/sbt-conductr/pull/227 because it actually makes sense to test the conductr-cli master branch. sbt-conductr heavily relies on the conductr-cli and we do not have integration tests in the conductr-cli project. So the sbt-conductr integration tests are also the integration tests of the conductr-cli.

Therefore it makes sense to use the latest commit of the master branch instead of the latest conductr-cli release.

This way, we will catch errors before releasing a new conductr-cli version.