Update conduct info to inspect bundle (#250)
* Update conduct info to inspect bundle

This is to update sbt-conductr to match with changes to `conduct info` as part of released version `1.1.0`.

From CLI version `1.1.0` the `conduct info` command accepts bundle id or bundle name as optional input argument, allowing inspection of bundle.

* Wait for the number of agent process to eventually settle to expected number

This is because the agent process launched by `sandbox run` is moved from its parent process to pid `1`. When attempt to assert the number of process while the move is taking place will yield failure in the test result.

* Sandbox end-to-end test: launching Visualizer and monitoring related bundles is no longer required

This is because `sandbox run` will always open the required port for all features during startup to allow feature bundles to be run after sandbox is started.

* sandbox-all-args test: launch lite-logging in place of the actual logging

This is because we don't perform any assertion on the actual bundle, we're just interested in seeing if sandbox will run with `-f` or `--feature` supplied.