Introduces tags
ConductR 2.1 tags are now used in place of compatibility version when producing packages. Tags are destined for human consumption, whereas compatibility versions are something that can help ConductR reason about whether a bundle with the same name is the same type of bundle e.g. when performing continuous delivery, it must be determined that a bundle can be replaced with another one. A default tag is generated from a project's version (compatibility versions also continue to be derived from a project's version).

Tag usage here is similar as per that found with Docker and git.

The README has also been updated for tags and slightly revised for an improvement in flow.